Redo Overwatch Path Warning logic
There were infrequent cases where it made the wrong prediction (given
a path that re-enters peripheral from a non-cover cell that was
blocked by tall cover the guard turns to face your current cell after
the move, but UITR predicted a turn to the previous cell based on a
different tracking case).

Rewritten to be more obvious that it should match the sim logic, with
separate blocks for the different basegame methods that perform
"possibly turn in response to the movement" and "react to
appearance/disappearance of the target".